JDP210601-41: implementation and changes in CartItem's relational methods and classes

+++ src/main/java/com/kodilla/ecommerce/dto/CartItemDto.java
@@ -3,13 +3,13 @@ package com.kodilla.ecommerce.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class CartItemDto {
 
     private Long id;
-
     private String name;
-
-    private Long quantity;
+    private BigDecimal quantity;
 }
